requirements.txt
# PyNaCl
six==1.10.0        --hash=sha512:a41b40b720c5267e4a47ffb98cdc79238831b4fbc0b20abb125504881b73ae38d5ef0215ee91f0d3582e7887244346e45da9410195d023105fccd96239f0ee95
pycparser==2.17    --hash=sha512:c9caaa8d256748e0623d077b11931abb38d19367136c70a835f7587e1f7ceb64f3acb7a983dcb68bedd2cf187517762a5753844e8ed58d1d9ed6f364c55839b4
cffi==1.9.1        --hash=sha512:0cff1ba30d40837b116c74cb5e1238f70aba40ecedf6034d0e5d7f9be2bb06b2aa9c105cb6ba19b50da967fe9cf5546a5de3255aeabf2c21b24c534fb0f6ed56
PyNaCl==1.1.1      --hash=sha512:6410f6ed2a474fefd5df425ea7e76fbe527a9d2ed09b36291caf2c5d0e68704e58caa694e06b01ea61323b2ef16ce85c1478191cf49d7eea969395a3d74d09a8

argon2==0.1.10     --hash=sha512:d9df98cff4b13cbcd36f5ecea7e3f62e7c21d2bda07d98cfe97f451af0ac3548ce3150c2c35f9f5a45e2143ecf552b6f4807c758ab7ac56a1940aeba3f144ba8
pyserial==3.3      --hash=sha512:19545d2121e0f43aba7e423df94c6836aba037b86500e2cd9b59555c80fb8596950329833b6d7aeb12a5d3b8e6639a14d64df984bf4e78a1f646364a05e08e4c
virtualenv==15.1.0 --hash=sha512:9988af801d9ad15c3f9831489ee9b49b54388e8349be201e7f7db3f2f1e59d033d3117f12e2f1909d65f052c5f1eacd87a894c6f7f703d770add3a0179e95863